requirements.txt
pandas==2.2.3
selenium==4.10.0
tqdm==4.66.1
python-dotenv==1.0.1
webdriver-manager==4.0.2
flask==3.0.0